+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+* Only redirect to NGINX when the correct content-type is set
+
 ## [3.6.2] - 2016-04-27
 ### Fixed
 * Don't gunzip data for clients that accept gzip when streaming